# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LocumMap Chennai API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /auth/otp/request:
    post:
      summary: Request OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
              example:
                phone: "+919900000001"
      responses:
        "200":
          description: OTP sent
  /auth/otp/verify:
    post:
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                otp:
                  type: string
                role:
                  type: string
              example:
                phone: "+919900000001"
                otp: "123456"
                role: "clinic_admin"
      responses:
        "200":
          description: Token
...